task 14 update1

+++ 9-binary_tree_height2.c
@@ -1,13 +1,13 @@
 #include "binary_trees.h"
 
-long long int max(long long int num1, long long int num2);
+long int max(long int num1, long int num2);
 
 /**
  * binary_tree_height2 - measures the height of a binary tree.
  * @tree: a pointer to the root node of the tree to measure the height.
  * Return: -1 if tree is NULL else height of tree.
  */
-long long int  binary_tree_height2(const binary_tree_t *tree)
+long int  binary_tree_height2(const binary_tree_t *tree)
 {
 	long long int left_height, right_height;
 
@@ -24,7 +24,7 @@ long long int  binary_tree_height2(const binary_tree_t *tree)
  * @num2: second number.
  * Return: the maximum number.
  */
-long long int max(long long int num1, long long int num2)
+long int max(long int num1, long int num2)
 {
 	if (num1 > num2)
 		return (num1);
